refactor(tools/galign): promote random_{unit,rotation} to internal API
